requirements.txt
joblib==0.17.0
pydantic==1.7.3
torch==1.7.1
loguru==0.5.3
transformers==2.5.1
pandas==0.25.0
tqdm==4.36.1
jieba==0.42.1
numpy==1.19.5
PyYAML==5.4.1
